Final Push
I have been working on this prototype for a very long time and learned a lot about the engine because of it. At this point in time, the code is quite garbled and confusing. While I feel somewhat sentimental about the project, I think I am better off learning through a new one. I will possibly revisit this when I gain more experience.